Update README.md
A Paired Device: The Passive Interface and Reader

The Passive Interface and Reader comprise a unique method to validate the identity of a sovereign individual, resulting in the creation of a DID, (Decentralized Identity). This DID, or proof of identity allows the individual to securely connect to the Cardano Blockchain and perform transactions. These transactions are of multiple types, including web3 transactions.

It is available in remote or rural areas, where there may be little or no internet service. It is not intended to replace biometric DID solutions, but merely offers an alternative that may be useful as the user-base of Cardano expands. It also provides a basis for empirical research to considering what we may gain by shifting our focus away from legacy screen-based methods of interface.

The Passive Interface

The Passive Interface in this iteration, can be defined as a tactile physical object that contains no active electronic circuitry to function. It can be worn or held, and is in possession of the User. It does not require the use of a cell phone or other networked connection.

For those familiar with blockchain, it is in a sense like a hardware wallet, which stores information needed to perform transactions. Only it is not an active device. It can not be altered through environmental changes. It has no batteries. It is not useful for the perpetrator if stolen. There is no seed phrase. It only has purpose for the owner of it—the physical human being, and it fits within their cultural system of reference—It is not meaningless to those who created and possess it. It is not designed to be disposable. It is like a fossil. Over time, its meaning has persistence.

If on blockchain transactions are immutable, (have an indelible persistence):

An Object, such as a Passive Interface becomes a parallel trace of immutable history.

The Reader The Passive Interface is paired with a unique DNA validation method, which is the Reader of the information stored within it. .

These two components, work together, to validate a user, and then provide access to connect to the blockchain, in remote or rural areas, where there may be little or no internet service. It is not intended to replace biometric identifier solutions, but merely to offer an alternative solution as the user-base of Cardano expands.

The Passive Interface/DNA validation pair provides a safe and secure DID which minimizes the exploitation of online ID validation schemes which can be used for other surveillance interests and violation of privacy, which is of great concern to indigenous communities and is becoming an expanding concern for the broader population. These online biometric methods are often seen as prohibitive when considering engaging on blockchain.